Changed Game Pages and Routing/Updated npm modules/Added Signup Duplicate Email Warning
