+++ senior_project_dea-main/package.json
@@ -5,19 +5,21 @@
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.2.2",
+    "@testing-library/user-event": "^14.4.3",
+    "bootstrap": "^5.2.3",
     "concurrently": "^7.6.0",
     "express-validator": "^6.14.3",
     "mdb-react-ui-kit": "^5.1.0",
+    "npm": "^9.4.2",
+    "npm-check": "^6.0.1",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.5.0",
+    "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
     "react-router-bootstrap": "^0.26.2",
-    "react-router-dom": "^6.4.2",
+    "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
     "watch": "^1.0.2",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
