+++ senior_project_dea-main/src/App.js
@@ -3,6 +3,8 @@ import MyNavbar from './components/Navbar';
 import MyWelcomePage from './components/Welcome';
 import LearnPage from './components/Learn';
 import GamePage from './components/Game';
+import GameTraditionalPage from './components/GameTraditional';
+import GameAdventurePage from './components/GameAdventure';
 import ProfilePage from './components/Profile';
 import { Routes, Route } from "react-router-dom"
 import Login from './components/Login'
@@ -14,7 +16,6 @@ import '../node_modules/bootstrap/dist/css/bootstrap.min.css'
 import { useLocation } from 'react-router-dom';
 import QuestionCRUD from './components/QuestionCRUD';
 import { useEffect } from 'react';
-
 
 function App() {
   
@@ -61,6 +62,8 @@ function App() {
               <Route path="/welcome" element={<MyWelcomePage />} />
               <Route path="/learn" element={<LearnPage />} />
               <Route path="/game" element={<GamePage />} />
+              <Route path="/gameTraditional" element={<GameTraditionalPage />} />
+              <Route path="/gameAdventure" element={<GameAdventurePage />} />
               <Route path="/myprofile" element={<ProfilePage />} />
               <Route exact path="/" element={<Login />} />
               <Route path="/sign-in" element={<Login />} />
