+++ senior_project_dea-main/src/components/GameAdventure.js
@@ -1,0 +1,9 @@
+function GameAdventurePage() {
+
+    //HTML Goes Here
+    //return (
+
+    //);
+}
+
+export default GameAdventurePage;
